feat(frontend): update app layout to show pages in the desired order
