+++ frontend/src/app/(protected)/(tabs)/(home)/_layout.jsx
@@ -3,7 +3,7 @@ import { Stack } from "expo-router";
 export default function Layout() {
   return (
     <Stack screenOptions={{ headerShown: false }}>
-      <Stack.Screen name="index" options={{ title: "Home" }} />
+      <Stack.Screen name="index" options={{ title: "CurbsideDropoff" }} />
     </Stack>
   );
 }
